+++ packages/editor/src/Editor.tsx
@@ -19,7 +19,7 @@ import TreeViewPlugin from "./plugins/TreeViewPlugin";
 import ListMaxIndentLevelPlugin from "./plugins/ListMaxIndentPlugin";
 import FloatingLinkEditorPlugin from "./plugins/FloatingLinkEditorPlugin";
 import { editorConfig } from "./config";
-import { LINE_BREAK_FIX } from "./transformers";
+import { COLLAPSIBLE, LINE_BREAK_FIX } from "./transformers";
 import { Language } from "./i18n";
 import { LocaleContext } from "./context/LocaleContext";
 
@@ -78,7 +78,7 @@ export const Editor = ({
           initialConfig={{
             ...editorConfig,
             editorState: () => {
-              $convertFromMarkdownString(content, [...TRANSFORMERS]);
+              $convertFromMarkdownString(content, [...TRANSFORMERS, COLLAPSIBLE]);
             },
           }}
         >
@@ -108,7 +108,11 @@ export const Editor = ({
             <OnChangePlugin
               onChange={(editorState) => {
                 editorState.read(() => {
-                  const markdown = $convertToMarkdownString([...TRANSFORMERS, LINE_BREAK_FIX]);
+                  const markdown = $convertToMarkdownString([
+                    ...TRANSFORMERS,
+                    COLLAPSIBLE,
+                    LINE_BREAK_FIX,
+                  ]);
                   onChange && onChange(markdown);
                 });
               }}