+++ packages/editor/src/config.ts
@@ -1,6 +1,11 @@
 import { HeadingNode, QuoteNode } from "@lexical/rich-text";
 import { LinkNode } from "@lexical/link";
 import { ListItemNode, ListNode } from "@lexical/list";
+import {
+  CollapsibleContainerNode,
+  CollapsibleContentNode,
+  CollapsibleTitleNode,
+} from "./nodes/CollapsibleNodes";
 
 export const editorConfig = {
   namespace: "FormBuilder",
@@ -15,5 +20,14 @@ export const editorConfig = {
     throw error;
   },
   // Any custom nodes go here
-  nodes: [HeadingNode, QuoteNode, LinkNode, ListItemNode, ListNode],
+  nodes: [
+    HeadingNode,
+    QuoteNode,
+    LinkNode,
+    ListItemNode,
+    ListNode,
+    CollapsibleContainerNode,
+    CollapsibleTitleNode,
+    CollapsibleContentNode,
+  ],
 };